Column selection
IMPORTANT: append the callback function in app.py and do not add a new one. Both will break :laughing:

Added a working column selector that is linked to the visualizations. We can now use a column that all visualizations share.

In the map, it is for now diplayed as additional text in the hover function.